refactor: move ETag generation logic to a separate utility file for improved code organization and reuse
